requirements.txt
# N8N Workflows API Dependencies
# Core API Framework
fastapi>=0.68.0
uvicorn>=0.15.0
pyyaml>=5.4.1
python-dotenv>=0.19.0
slowapi>=0.1.4
typing-extensions>=4.0.0